requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask-migrate==4.0.7
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3
gunicorn==22.0.0
pytest==8.3.2
